endpoint: move `TestReadEpsFromDirNames` to `pkg/endpoint`
The test was previously in `daemon`, and necessitated leaking a lot of
Endpoint-related state. Instead, refactor the test into the Endpoint package.
Do not regenerate the endpoints needed for testing; instead, write their
headerfiles using `writeHeaderfile`, since that is the operation which
is needed to be performed in order to perform the test to read the serialized
Endpoints from their corresponding headerfiles.

Signed-off by: Ian Vernon <ian@cilium.io>
